fix: 상품 단위/소수점 전수 점검 + BOM 소수점 지원
## 점검 결과
- stock_ledger.qty: NUMERIC (소수점 지원 확인 ✅)
- 완제품 단위=개: safe_qty()가 정수 강제 → 정상 (완제품은 정수)
- 원료 단위=kg/g: safe_qty()가 소수점 보존 → 정상
- 생산/소분 UI: step="any" → 소수점 입력 가능
- 수율 계산: float 처리 → 정상

## 수정
- bom_cost_service.py: BOM 파싱 int() → float() (소수점 BOM 지원)
- 완제품 96건 단위 g→개 일괄 수정 (DB 직접)
- 상품관리 기본 필터: 완제품만 표시

+++ services/bom_cost_service.py
@@ -21,7 +21,7 @@ def parse_components(comp_str):
         if idx > 0:
             name = item[:idx].strip()
             try:
-                qty = int(item[idx + 1:].strip())
+                qty = float(item[idx + 1:].strip())
             except ValueError:
                 name = item
                 qty = 1
